+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 9.3
 - Handle case where Course Period Parent ID is null in Courses.php
 - SQL order by Period title in Periods.php, DailySummary.php & StudentSummary.php
 - SQL get Period title if no short name set in AddAbsences.php
+- Use DBLastInsertID() instead of DBSeqNextID() in Moodle/includes/ImportUsers.fnc.php
 
 Changes in 9.2.2
 ----------------